added tests around calculating factors
